requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
jinja2==3.1.4
python-dotenv==1.0.1
anthropic>=0.40.0
boto3>=1.34.0
bcrypt>=4.0.1
itsdangerous>=2.2.0
python-multipart>=0.0.9
